requirements.txt
pillow
pyzbar
base45
cbor2
